Changes to unify storage reader #3024
