+++ plaso/storage/file_interface.py
@@ -104,161 +104,3 @@ class BaseStorageFile(interface.BaseStore):
     """
     self._attribute_container_sequence_numbers[
         container_type] = next_sequence_number
-
-
-class StorageFileReader(interface.StorageReader):
-  """File-based storage reader interface."""
-
-  def __init__(self, path):
-    """Initializes a storage reader.
-
-    Args:
-      path (str): path to the input file.
-    """
-    super(StorageFileReader, self).__init__()
-    self._path = path
-    self._storage_file = None
-
-  def Close(self):
-    """Closes the storage reader."""
-    if self._storage_file:
-      self._storage_file.Close()
-      self._storage_file = None
-
-  def GetAttributeContainerByIdentifier(self, container_type, identifier):
-    """Retrieves a specific type of container with a specific identifier.
-
-    Args:
-      container_type (str): container type.
-      identifier (AttributeContainerIdentifier): attribute container identifier.
-
-    Returns:
-      AttributeContainer: attribute container or None if not available.
-    """
-    return self._storage_file.GetAttributeContainerByIdentifier(
-        container_type, identifier)
-
-  def GetAttributeContainers(self, container_type):
-    """Retrieves a specific type of attribute containers.
-
-    Args:
-      container_type (str): attribute container type.
-
-    Returns:
-      generator(AttributeContainers): attribute container generator.
-    """
-    return self._storage_file.GetAttributeContainers(container_type)
-
-  def GetFormatVersion(self):
-    """Retrieves the format version of the underlying storage file.
-
-    Returns:
-      int: the format version, or None if not available.
-    """
-    if self._storage_file:
-      return self._storage_file.format_version
-
-    return None
-
-  def GetNumberOfAnalysisReports(self):
-    """Retrieves the number analysis reports.
-
-    Returns:
-      int: number of analysis reports.
-    """
-    return self._storage_file.GetNumberOfAttributeContainers(
-        self._CONTAINER_TYPE_ANALYSIS_REPORT)
-
-  def GetNumberOfEventSources(self):
-    """Retrieves the number of event sources.
-
-    Returns:
-      int: number of event sources.
-    """
-    return self._storage_file.GetNumberOfAttributeContainers(
-        self._CONTAINER_TYPE_EVENT_SOURCE)
-
-  def GetSerializationFormat(self):
-    """Retrieves the serialization format of the underlying storage file.
-
-    Returns:
-      str: the serialization format, or None if not available.
-    """
-    if self._storage_file:
-      return self._storage_file.serialization_format
-
-    return None
-
-  def GetSessions(self):
-    """Retrieves the sessions.
-
-    Returns:
-      generator(Session): session generator.
-    """
-    return self._storage_file.GetSessions()
-
-  def GetSortedEvents(self, time_range=None):
-    """Retrieves the events in increasing chronological order.
-
-    This includes all events written to the storage including those pending
-    being flushed (written) to the storage.
-
-    Args:
-      time_range (Optional[TimeRange]): time range used to filter events
-          that fall in a specific period.
-
-    Returns:
-      generator(EventObject): event generator.
-    """
-    return self._storage_file.GetSortedEvents(time_range=time_range)
-
-  def GetStorageType(self):
-    """Retrieves the storage type of the underlying storage file.
-
-    Returns:
-      str: the storage type, or None if not available.
-    """
-    if self._storage_file:
-      return self._storage_file.storage_type
-
-    return None
-
-  def HasAttributeContainers(self, container_type):
-    """Determines if a store contains a specific type of attribute container.
-
-    Args:
-      container_type (str): attribute container type.
-
-    Returns:
-      bool: True if the store contains the specified type of attribute
-          containers.
-    """
-    return self._storage_file.HasAttributeContainers(container_type)
-
-  # TODO: remove, this method is kept for backwards compatibility reasons.
-  def ReadSystemConfiguration(self, knowledge_base):
-    """Reads system configuration information.
-
-    The system configuration contains information about various system specific
-    configuration data, for example the user accounts.
-
-    Args:
-      knowledge_base (KnowledgeBase): is used to store the system configuration.
-    """
-    self._storage_file.ReadSystemConfiguration(knowledge_base)
-
-  def SetSerializersProfiler(self, serializers_profiler):
-    """Sets the serializers profiler.
-
-    Args:
-      serializers_profiler (SerializersProfiler): serializers profiler.
-    """
-    self._storage_file.SetSerializersProfiler(serializers_profiler)
-
-  def SetStorageProfiler(self, storage_profiler):
-    """Sets the storage profiler.
-
-    Args:
-      storage_profiler (StorageProfiler): storage profiler.
-    """
-    self._storage_file.SetStorageProfiler(storage_profiler)
