+++ plaso/storage/interface.py
@@ -481,137 +481,3 @@ class BaseStore(object):
       raise IOError('Task completion not supported by storage type.')
 
     self._WriteNewAttributeContainer(task_start)
-
-
-# pylint: disable=redundant-returns-doc,redundant-yields-doc
-class StorageReader(object):
-  """Storage reader interface."""
-
-  _CONTAINER_TYPE_ANALYSIS_REPORT = reports.AnalysisReport.CONTAINER_TYPE
-  _CONTAINER_TYPE_ANALYSIS_WARNING = warnings.AnalysisWarning.CONTAINER_TYPE
-  _CONTAINER_TYPE_EVENT = events.EventObject.CONTAINER_TYPE
-  _CONTAINER_TYPE_EVENT_DATA = events.EventData.CONTAINER_TYPE
-  _CONTAINER_TYPE_EVENT_DATA_STREAM = events.EventDataStream.CONTAINER_TYPE
-  _CONTAINER_TYPE_EVENT_SOURCE = event_sources.EventSource.CONTAINER_TYPE
-  _CONTAINER_TYPE_EVENT_TAG = events.EventTag.CONTAINER_TYPE
-  _CONTAINER_TYPE_EXTRACTION_WARNING = warnings.ExtractionWarning.CONTAINER_TYPE
-  _CONTAINER_TYPE_PREPROCESSING_WARNING = (
-      warnings.PreprocessingWarning.CONTAINER_TYPE)
-  _CONTAINER_TYPE_RECOVERY_WARNING = warnings.RecoveryWarning.CONTAINER_TYPE
-  _CONTAINER_TYPE_TASK_COMPLETION = tasks.TaskCompletion.CONTAINER_TYPE
-  _CONTAINER_TYPE_TASK_START = tasks.TaskStart.CONTAINER_TYPE
-
-  def __enter__(self):
-    """Make usable with "with" statement."""
-    return self
-
-  # pylint: disable=unused-argument
-  def __exit__(self, exception_type, value, traceback):
-    """Make usable with "with" statement."""
-    self.Close()
-
-  @abc.abstractmethod
-  def Close(self):
-    """Closes the storage reader."""
-
-  @abc.abstractmethod
-  def GetAttributeContainerByIdentifier(self, container_type, identifier):
-    """Retrieves a specific type of container with a specific identifier.
-
-    Args:
-      container_type (str): container type.
-      identifier (AttributeContainerIdentifier): attribute container identifier.
-
-    Returns:
-      AttributeContainer: attribute container or None if not available.
-    """
-
-  @abc.abstractmethod
-  def GetAttributeContainers(self, container_type):
-    """Retrieves a specific type of attribute containers.
-
-    Args:
-      container_type (str): attribute container type.
-
-    Returns:
-      generator(AttributeContainers): attribute container generator.
-    """
-
-  @abc.abstractmethod
-  def GetNumberOfAnalysisReports(self):
-    """Retrieves the number analysis reports.
-
-    Returns:
-      int: number of analysis reports.
-    """
-
-  @abc.abstractmethod
-  def GetNumberOfEventSources(self):
-    """Retrieves the number of event sources.
-
-    Returns:
-      int: number of event sources.
-    """
-
-  @abc.abstractmethod
-  def GetSessions(self):
-    """Retrieves the sessions.
-
-    Yields:
-      Session: session.
-    """
-
-  @abc.abstractmethod
-  def GetSortedEvents(self, time_range=None):
-    """Retrieves the events in increasing chronological order.
-
-    This includes all events written to the storage including those pending
-    being flushed (written) to the storage.
-
-    Args:
-      time_range (Optional[TimeRange]): time range used to filter events
-          that fall in a specific period.
-
-    Yields:
-      EventObject: event.
-    """
-
-  @abc.abstractmethod
-  def HasAttributeContainers(self, container_type):
-    """Determines if a store contains a specific type of attribute container.
-
-    Args:
-      container_type (str): attribute container type.
-
-    Returns:
-      bool: True if the store contains the specified type of attribute
-          containers.
-    """
-
-  # TODO: remove, this method is kept for backwards compatibility reasons.
-  @abc.abstractmethod
-  def ReadSystemConfiguration(self, knowledge_base):
-    """Reads system configuration information.
-
-    The system configuration contains information about various system specific
-    configuration data, for example the user accounts.
-
-    Args:
-      knowledge_base (KnowledgeBase): is used to store the system configuration.
-    """
-
-  @abc.abstractmethod
-  def SetSerializersProfiler(self, serializers_profiler):
-    """Sets the serializers profiler.
-
-    Args:
-      serializers_profiler (SerializersProfiler): serializers profiler.
-    """
-
-  @abc.abstractmethod
-  def SetStorageProfiler(self, storage_profiler):
-    """Sets the storage profiler.
-
-    Args:
-      storage_profiler (StorageProfiler): storage profile.
-    """
