+++ plaso/storage/redis/writer.py
@@ -75,24 +75,6 @@ class RedisStorageWriter(writer.StorageWriter):
         redis_client=redis_client, session_identifier=session_identifier,
         task_identifier=task_identifier)
 
-  def ReadSystemConfiguration(self, knowledge_base):
-    """Reads system configuration information.
-
-    The system configuration contains information about various system specific
-    configuration data, for example the user accounts.
-
-    Args:
-      knowledge_base (KnowledgeBase): is used to store the preprocessing
-          information.
-
-    Raises:
-      IOError: always, as the Redis store does not support preprocessing
-          information.
-      OSError: always, as the Redis store does not support preprocessing
-          information.
-    """
-    raise IOError('Preprocessing information not supported by the redis store.')
-
   def WritePreprocessingInformation(self, knowledge_base):
     """Writes preprocessing information.
