+++ plaso/storage/sqlite/reader.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 """Reader for SQLite storage files."""
 
-from plaso.storage import file_interface
+from plaso.storage import reader
 from plaso.storage.sqlite import sqlite_file
 
 
-class SQLiteStorageFileReader(file_interface.StorageFileReader):
+class SQLiteStorageFileReader(reader.StorageReader):
   """SQLite-based storage file reader."""
 
   def __init__(self, path):
@@ -14,6 +14,40 @@ class SQLiteStorageFileReader(file_interface.StorageFileReader):
     Args:
       path (str): path to the input file.
     """
-    super(SQLiteStorageFileReader, self).__init__(path)
-    self._storage_file = sqlite_file.SQLiteStorageFile()
-    self._storage_file.Open(path=path)
+    super(SQLiteStorageFileReader, self).__init__()
+    self._path = path
+    self._store = sqlite_file.SQLiteStorageFile()
+    self._store.Open(path=path)
+
+  def GetFormatVersion(self):
+    """Retrieves the format version of the underlying storage file.
+
+    Returns:
+      int: the format version, or None if not available.
+    """
+    if self._store:
+      return self._store.format_version
+
+    return None
+
+  def GetSerializationFormat(self):
+    """Retrieves the serialization format of the underlying storage file.
+
+    Returns:
+      str: the serialization format, or None if not available.
+    """
+    if self._store:
+      return self._store.serialization_format
+
+    return None
+
+  def GetStorageType(self):
+    """Retrieves the storage type of the underlying storage file.
+
+    Returns:
+      str: the storage type, or None if not available.
+    """
+    if self._store:
+      return self._store.storage_type
+
+    return None
